+++ R/convert.R
@@ -57,7 +57,7 @@ convertMSFilesOpenMS <- function(inFiles, outFiles, to, extraOpts, logPath, maxP
 #' are supported for input files.
 #'
 #' @param files A \code{character} vector with input files. Alternatively, if
-#'   \code{dir=TRUE}, then a \code{character} vector with one or more
+#'   \code{dirs=TRUE}, then a \code{character} vector with one or more
 #'   directories from which input files are automatically selected.
 #' @param outPath A character vector specifying directories that should be used
 #'   for the output. Will be re-cycled if necessary. If \code{NULL}, output