+++ NEWS.md
@@ -1,6 +1,8 @@
 # patRoon 0.1.0.9000
 
 ## November 2018
+* Added `addAllDAEICs()` function.
+* Renamed `mzWidth` argument of `addDAEIC()` to `mzWindow`.
 * Normalization of compound scores: normalization method can now be set and specified scorings can be excluded.
 * Store/report IUPACName (as compoundName) from MetFrag PubChem data.
 * Renamed trivialName to compoundName for compound tables.